requirements.txt
fastapi
uvicorn[standard]
pydantic
easyocr
torch
transformers
faiss-cpu
Pillow
rembg
opencv-python-headless
requests
python-multipart
pandas
paddleocr
ultralytics
Pillow
langchain
langgraph
langchain_groq
typing_extensions
sentence_transformers
rank-bm25